Optimise schema creation where SQLite is used

+++ features/Context/AliceFixturesContext.php
@@ -10,13 +10,15 @@ use Behat\Gherkin\Node\PyStringNode;
 use Behat\Behat\Event\ScenarioEvent;
 use Behat\MinkExtension\Context\MinkContext;
 use Behat\Symfony2Extension\Context\KernelAwareInterface;
-use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\DBAL\Platforms\SqlitePlatform;
 use Doctrine\ORM\Tools\SchemaTool;
 use Symfony\Component\HttpKernel\KernelInterface;
 use Nelmio\Alice\LoaderInterface;
 use Nelmio\Alice\ORMInterface;
 use Nelmio\Alice\Loader\Base as Loader;
 use Nelmio\Alice\ORM\Doctrine as Persister;
+use Doctrine\DBAL\Driver\Connection;
+use Doctrine\ORM\EntityManager;
 
 
 class AliceFixturesContext extends BehatContext implements FixturesContext, KernelAwareInterface
@@ -75,14 +77,65 @@ class AliceFixturesContext extends BehatContext implements FixturesContext, Kern
      */
     public function buildSchema()
     {
-        foreach ($this->getEntityManagers() as $entityManager) {
-            $metadata = $this->getMetadata($entityManager);
+        foreach ($this->getEntityManagers() as $name => $entityManager) {
+            $connection = $entityManager->getConnection();
+            $dbPlatform = $connection->getDatabasePlatform();
 
+            if ($dbPlatform->getName() == 'sqlite') {
+                $this->prepareEmptySqliteSchema($name, $entityManager, $connection, $dbPlatform);
+                continue;
+            }
+
+            $metadata = $this->getMetadata($entityManager);
             if (!empty($metadata)) {
                 $tool = new SchemaTool($entityManager);
                 $tool->dropSchema($metadata);
                 $tool->createSchema($metadata);
             }
+        }
+    }
+
+    /**
+     * @param string $entityManagerName
+     * @param EntityManager $entityManager
+     * @param Connection $connection
+     * @param SqlitePlatform $dbPlatform
+     */
+    private function prepareEmptySqliteSchema($entityManagerName, $entityManager, $connection, SqlitePlatform $dbPlatform)
+    {
+        static $schemaCreated = [];
+        $metadata = $this->getMetadata($entityManager);
+
+        if (!isset($schemaCreated[$entityManagerName])) {
+            $tool = new SchemaTool($entityManager);
+            $tool->dropSchema($metadata);
+            $tool->createSchema($metadata);
+            $schemaCreated[$entityManagerName] = true;
+            return;
+        }
+
+        /*
+         * If the schema has already been created it's much faster to delete all of the data than to drop and recreate
+         * the schema.
+         */
+        foreach ($metadata as $classMetaData) {
+            try {
+                $connection->beginTransaction();
+                $connection->query('PRAGMA foreign_keys = OFF');
+                $connection->executeUpdate('DELETE FROM '.$dbPlatform->quoteIdentifier($classMetaData->getTableName()));
+                $connection->query('PRAGMA foreign_keys = ON');
+                $connection->commit();
+            }
+            catch (\Exception $e) {
+                $connection->rollback();
+            }
+        }
+
+        try {
+            $connection->executeUpdate('DELETE FROM sqlite_sequence');
+        }
+        catch (\Exception $e) {
+            //No sqlite_sequence table exists - that's ok, it just means we have no auto incrementing columns
         }
     }
 
@@ -101,7 +154,7 @@ class AliceFixturesContext extends BehatContext implements FixturesContext, Kern
     }
 
     /**
-     * @param EntityManagerInterface $entityManager
+     * @param EntityManager $entityManager
      *
      * @return array
      */
@@ -111,7 +164,7 @@ class AliceFixturesContext extends BehatContext implements FixturesContext, Kern
     }
 
     /**
-     * @return EntityManagerInterface[]
+     * @return EntityManager[]
      */
     protected function getEntityManagers()
     {
